requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.11,<3.0
pydantic-settings>=2.10,<3.0
python-multipart>=0.0.20,<1.0
PyJWT>=2.10,<3.0
email-validator>=2.2,<3.0
pytest>=8.4,<9.0
httpx>=0.28,<1.0
